chore: moving kafka client
