+++ kafkaclient/client.go
