+++ kafkaclient/logger.go
