+++ kafkaclient/producer.go
