+++ kafkaclient/testdata/keystore/kafka.keystore.jks
