+++ kafkaclient/testdata/ssh/test_key.pub
