+++ kafkaclient/testdata/truststore/kafka.truststore.jks
