+++ kafkaclient/testdata/user1.avsc
