+++ kafkaclient/testdata/user2.avsc
