+++ kafkaclient/testutil/util.go
